+++ lib/insite/page/undefined_page.rb
@@ -33,6 +33,14 @@ module Insite
   EOF
     end
 
+    def driver
+      @browser.driver
+    end
+
+    def html
+      @browser.html
+    end
+
     def initialize(site)
       @site    = site
       @browser = process_browser
